build: remove gcc_version from common.gypi
The gcc_version variable selectively enabled -fvisibility=hidden, a flag
that is available as of gcc 4.0.  Because libuv only supports gcc 4.2 or
newer, there is no reason to keep it around.

If an embedder insists on building libuv with an older compiler, they
can do so by overriding the flag in their build.gyp with the following
snippet:

    'cflags!': ['-fvisibility=hidden']

But caveat emptor, it's unlikely that libuv actually builds with a
compiler that old.