linux: store inotify watchers in a hash table
When there are lots of uv_fs_event_t file watchers on a busy file tree,
say on the order of 50,000-100,000 watchers and hundreds or thousands
events per second, the overhead of binary tree search/insertion/removal
starts getting noticeable.

Switch to a hash table. There is some overhead involved in growing the
hash table but once it reaches a steady state, lookups are O(1) in a
cache-friendlier way than the pointer chasing of the binary tree.

Fixes: https://github.com/libuv/libuv/issues/3780